move common cache functions to a separate module
